+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-CHANGES in 1.0.0
+Changes in 1.0.0
 ================
 
-* Imported from origin repo
+* Initial port from private repository
